# docs/swagger.yaml
basePath: /
definitions:
...
  talacert-api_internal_dto.CreateUserRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      role:
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  talacert-api_internal_dto.UpdateUserRequest:
    properties:
      email:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  talacert-api_internal_dto.UserChangePasswordRequest:
    properties:
      password:
        minLength: 8
        type: string
    required:
    - password
    type: object
  talacert-api_internal_dto.UserChangeRoleRequest:
    properties:
      role:
        type: string
    required:
    - role
    type: object
...
host: 127.0.0.1:8080
info:
  contact: {}
  description: |-
    API de vérification de documents officiels.
    TalaCert permet de créer, gérer et vérifier l'authenticité des certificats, diplômes et attestations.
  title: TalaCert API
  version: "1.0"
paths:
...
  /api/v1/users:
    get:
      description: Retrieves all users.
      produces:
      - application/json
      responses:
        "200":
          description: Users retrieved successfully
          schema: {}
        "500":
          description: Failed to retrieve users
          schema: {}
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Creates a new user.
      parameters:
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/talacert-api_internal_dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema: {}
        "400":
          description: Invalid request payload
          schema: {}
        "409":
          description: User already exists
          schema: {}
        "500":
          description: Failed to create user
          schema: {}
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - Users
  /api/v1/users/{id}:
    delete:
      description: Deletes a user by their unique identifier.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema: {}
        "400":
          description: Invalid user ID
          schema: {}
        "404":
          description: User not found
          schema: {}
        "500":
          description: Failed to delete user
          schema: {}
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - Users
    get:
      description: Retrieves a user by their unique identifier.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User retrieved successfully
          schema: {}
        "400":
          description: Invalid user ID
          schema: {}
        "404":
          description: User not found
          schema: {}
        "500":
          description: Failed to retrieve user
          schema: {}
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Updates an existing user.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/talacert-api_internal_dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema: {}
        "400":
          description: Invalid user ID or request payload
          schema: {}
        "404":
          description: User not found
          schema: {}
        "409":
          description: User already exists
          schema: {}
        "500":
          description: Failed to update user
          schema: {}
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - Users
  /api/v1/users/{id}/password:
    patch:
      consumes:
      - application/json
      description: Changes the password of an existing user.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: New password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/talacert-api_internal_dto.UserChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed successfully
          schema: {}
        "400":
          description: Invalid user ID or request payload
          schema: {}
        "500":
          description: Failed to change password
          schema: {}
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - Users
  /api/v1/users/{id}/role:
    patch:
      consumes:
      - application/json
      description: Changes the role of an existing user.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: New user role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/talacert-api_internal_dto.UserChangeRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Role changed successfully
          schema: {}
        "400":
          description: Invalid user ID or request payload
          schema: {}
        "500":
          description: Failed to change role
          schema: {}
      security:
      - BearerAuth: []
      summary: Change user role
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"